Implement lunging
An experimental way to discourage long-distance luring or so-called
'pillar dancing', which tends to be rather repetitive and not very
fun. Gives monsters a 1/6 chance to get a free attack when moving
adjacent to their foe, 'lunging' at them. This should have a minimal
impact on most normal fights and even allow repositioning across a
reasonable distance with adjacent enemies.

If this proves too vicious, we can experiment with various compensation
- for example, we could increase player HP slightly, especially in the
early game.